Remove debouncer
